fix Too few arguments to function Laudis\Neo4j\Types\Node::__construct()
fix Too few arguments to function Laudis\Neo4j\Types\Node::__construct()

+++ src/Eloquent/Relations/HasOneOrMany.php
@@ -528,7 +528,9 @@ abstract class HasOneOrMany extends Relation implements RelationInterface
         $results = [];
         foreach ($id as $model) {
             $edge = $this->edge($model);
-            $results[] = $edge->delete();
+            if ($edge) {
+                $results[] = $edge->delete();
+            }
         }
 
         if ($touch) {
